support route mode: no change for routing

+++ files/client_down.sh
@@ -20,26 +20,29 @@ loger() {
 # Get uci setting
 route_mode=$(uci get minivtun.@minivtun[-1].route_mode_save 2>/dev/null)
 
-# Turn off NAT over VPN
-iptables -t nat -D POSTROUTING -o $intf -j MASQUERADE
-iptables -D FORWARD -i $intf -m state --state RELATED,ESTABLISHED -j ACCEPT
-iptables -D FORWARD -o $intf -j ACCEPT
-loger notice "Turn off NAT over $intf"
+# mode 3 is "no change for route"
+if [ "$route_mode" != 3 ]; then
+	# Turn off NAT over VPN
+	iptables -t nat -D POSTROUTING -o $intf -j MASQUERADE
+	iptables -D FORWARD -i $intf -m state --state RELATED,ESTABLISHED -j ACCEPT
+	iptables -D FORWARD -o $intf -j ACCEPT
+	loger notice "Turn off NAT over $intf"
 
-# Change routing table
-ip route del $server
-if [ "$route_mode" != 2 ]; then
-	ip route del 0.0.0.0/1
-	ip route del 128.0.0.0/1
-	loger notice "Default route changed to original route"
+	# Change routing table
+	ip route del $server
+	if [ "$route_mode" != 2 ]; then
+		ip route del 0.0.0.0/1
+		ip route del 128.0.0.0/1
+		loger notice "Default route changed to original route"
+	fi
+
+	# Remove route rules
+	if [ -f /tmp/minivtun_routes ]; then
+		sed -e "s/^/route del /" /tmp/minivtun_routes | ip -batch -
+		loger notice "Route rules have been removed"
+	fi
+
+	rm -rf /tmp/minivtun_routes
 fi
-
-# Remove route rules
-if [ -f /tmp/minivtun_routes ]; then
-	sed -e "s/^/route del /" /tmp/minivtun_routes | ip -batch -
-	loger notice "Route rules have been removed"
-fi
-
-rm -rf /tmp/minivtun_routes
 
 loger info "Script $0 completed"
